Fix parallel runtime CI boundaries
