+++ clients/desktop/package.json
@@ -8,6 +8,7 @@
     "format": "prettier --write 'src/**/*.{ts,tsx,js,jsx,json,css,scss,md}'",
     "format:wails": "prettier --write 'wailsjs/**/*.{ts,js}'",
     "dev": "vite",
+    "dev:worktree": "node scripts/dev-desktop.mjs",
     "build": "node --max-old-space-size=8192 ../../node_modules/typescript/bin/tsc && node --max-old-space-size=8192 ../../node_modules/vite/bin/vite.js build",
     "preview": "vite preview",
     "test": "vitest --config vitest.config.mts",
