+++ clients/desktop/vitest.config.mts
@@ -19,6 +19,9 @@ export default defineConfig({
   },
   test: {
     environment: 'node',
-    include: ['src/**/*.{test,spec}.{ts,tsx}'],
+    include: [
+      'src/**/*.{test,spec}.{ts,tsx}',
+      'scripts/**/*.{test,spec}.{js,mjs}',
+    ],
   },
 })
