+++ package.json
@@ -8,8 +8,8 @@
   "scripts": {
     "lint": "eslint --cache --cache-location node_modules/.cache/eslint/ 'clients/desktop/src/**/*.{js,jsx,ts,tsx}' 'clients/extension/src/**/*.{js,jsx,ts,tsx}' 'clients/desktop/vite.config.mts' 'clients/extension/vite.config.ts' 'core/**/*.{js,jsx,ts,tsx}' 'lib/**/*.{js,jsx,ts,tsx}'",
     "lint:fix": "eslint --cache --cache-location node_modules/.cache/eslint/ 'clients/desktop/src/**/*.{js,jsx,ts,tsx}' 'clients/extension/src/**/*.{js,jsx,ts,tsx}' 'clients/desktop/vite.config.mts' 'clients/extension/vite.config.ts' 'core/**/*.{js,jsx,ts,tsx}' 'lib/**/*.{js,jsx,ts,tsx}' --fix",
-    "dev:desktop": "node scripts/dev-desktop.mjs",
-    "dev:desktop:full": "node scripts/dev-desktop.mjs --full",
+    "dev:desktop": "yarn workspace @clients/desktop dev:worktree",
+    "dev:desktop:full": "yarn workspace @clients/desktop dev:worktree --full",
     "build:desktop": "NODE_OPTIONS=--max-old-space-size=8192 yarn workspace @clients/desktop build",
     "build:extension": "yarn workspace @clients/extension build",
     "build:extension:station": "yarn workspace @clients/extension build:station",
