+++ vitest.config.mts
@@ -14,7 +14,6 @@ export default defineConfig({
     include: [
       'core/**/*.{test,spec}.{js,ts,tsx}',
       'lib/**/*.{test,spec}.{js,ts,tsx}',
-      'scripts/**/*.{test,spec}.{js,mjs,ts}',
     ],
     exclude: ['**/node_modules/**', '**/dist/**', '**/tests/**', 'clients/**'],
   },
